Model, Controller e Migration adicionados para Comment, Subtask e Image
Migrações antigas tiveram atualizações em suas colunas também

+++ database/migrations/2022_02_15_195943_create_tasks_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
 				->on('categories');
 			$table->string('description');
 			$table->string('reference');
-			$table->integer('position');
+			$table->smallInteger('position');
 			$table->tinyText('color');
             $table->timestamps();
         });
